requirements.txt
numpy
pandas
pyarrow
duckdb
matplotlib
seaborn
torch
datasets
scikit-learn
xgboost
lightgbm
catboost
pypots
ray[serve]
chronos-forecasting
pyyaml
tqdm
streamlit
requests
fastapi
